colocate type definitions
