+++ packages/react-aria-widgets/src/Accordion/AccordionHeader.tsx
@@ -9,7 +9,28 @@ import useAccordionContext from 'src/Accordion/useAccordionContext';
 import useAccordionItemContext from 'src/Accordion/useAccordionItemContext';
 
 //Types
-import type { AccordionHeaderProps } from 'src/Accordion/types';
+import type {
+  AccordionRenderFunction,
+  AccordionRenderClass,
+  AccordionRenderStyle,
+} from 'src/Accordion/types';
+import type { BaseHeaderProps, BaseButtonProps } from 'src/Accordion/BaseAccordionHeader';
+
+export type HeaderProps = {
+  className?: string | AccordionRenderClass;
+  style?: React.CSSProperties | AccordionRenderStyle;
+} & Omit<BaseHeaderProps, 'className' | 'style'>;
+
+export type ButtonProps = {
+  className?: string | AccordionRenderClass;
+  style?: React.CSSProperties | AccordionRenderStyle;
+} & Omit<BaseButtonProps, 'className' | 'style'>;
+
+export interface AccordionHeaderProps {
+  children?: React.ReactNode | AccordionRenderFunction;
+  headerProps?: HeaderProps;
+  buttonProps?: ButtonProps;
+}
 
 function AccordionHeader({
   children = null,
