+++ packages/react-aria-widgets/src/Accordion/AccordionItem.tsx
@@ -4,8 +4,9 @@ import PropTypes from 'prop-types';
 //Contexts
 import { AccordionItemProvider } from 'src/Accordion/AccordionItemContext';
 
-//Types
-import type { AccordionItemProps } from 'src/Accordion/types';
+export type AccordionItemProps = React.PropsWithChildren<{
+  id: string;
+}>;
 
 function AccordionItem({
   children = null,
