+++ packages/react-aria-widgets/src/Accordion/useAccordion.ts
@@ -1,25 +1,43 @@
 import { useState, useCallback, useRef, useMemo, useEffect } from 'react';
 
 //Types
-import type {
-  UseAccordion,
-  ExpandedItems,
-  DisabledItems,
-  HeaderRef,
-  GetIsExpanded,
-  GetIsDisabled,
-  ToggleExpanded,
-  ToggleDisabled,
-  PushHeaderRef,
-  FocusHeaderIndex,
-  FocusHeaderId,
-  FocusPrevHeader,
-  FocusNextHeader,
-  FocusFirstHeader,
-  FocusLastHeader,
-  OnToggleExpanded,
-  OnToggleDisabled,
-} from 'src/Accordion/types';
+import type { ValidHTMLHeaderLevels } from 'src/utils/types';
+
+export type ExpandedItems = Set<string>;
+export type DisabledItems = Set<string>;
+
+export type AccordionHeaderButtonElement = HTMLButtonElement | HTMLElement | null;
+
+export interface AccordionHeaderRef {
+  elem: AccordionHeaderButtonElement;
+  id: string;
+}
+
+export type GetIsExpanded = (id: string) => boolean;
+export type GetIsDisabled = (id: string) => boolean;
+export type ToggleExpanded = (id: string) => void;
+export type ToggleDisabled = (id: string) => void;
+export type PushHeaderRef = (elem: AccordionHeaderButtonElement, id: string) => void;
+export type FocusHeaderIndex = (index: number) => void;
+export type FocusHeaderId = (id: string) => void;
+export type FocusPrevHeader = (id: string) => void;
+export type FocusNextHeader = (id: string) => void;
+export type FocusFirstHeader = () => void;
+export type FocusLastHeader = () => void;
+export type OnToggleExpanded = (expandedItems: ExpandedItems) => void;
+export type OnToggleDisabled = (disabledItems: DisabledItems) => void;
+export type OnFocusChange = ({ elem, index, id }: { elem: AccordionHeaderButtonElement; index: number; id: string }) => void;
+
+export interface UseAccordion {
+  allowMultiple?: boolean;
+  allowCollapseLast?: boolean;
+  headerLevel: ValidHTMLHeaderLevels;
+  initialExpanded?: string[];
+  initialDisabled?: string[];
+  onToggleExpanded?: OnToggleExpanded | undefined;
+  onToggleDisabled?: OnToggleDisabled | undefined;
+  onFocusChange?: OnFocusChange | undefined;
+}
 
 function _getIsExpanded(id: string, expandedItems: Set<string>) {
   return expandedItems.has(id);
@@ -51,7 +69,7 @@ export default function useAccordion({
 
   const [ expandedItems, setExpandedItems ] = useState<ExpandedItems>(_initialExpanded);
   const [ disabledItems, setDisabledItems ] = useState<DisabledItems>(_initialDisabled);
-  const headerRefs = useRef<HeaderRef[]>([]);
+  const headerRefs = useRef<AccordionHeaderRef[]>([]);
   const idToIndexMap = useRef<Map<string, number>>(new Map());
   const onToggleExpandedRef = useRef<OnToggleExpanded | null | undefined>(null);
   const onToggleDisabledRef = useRef<OnToggleDisabled | null | undefined>(null);
